requirements.txt
celery==5.2.0
dj-config-url
dj-rest-auth
Django
django-filter
django-templated-mail
djangorestframework
djangorestframework-simplejwt
djoser
drf-yasg
email-validator
gunicorn
psycopg2-binary
PyJWT
python-decouple
python3-openid
redis
requests
stripe
utils
pillow

